rename module for string transform processors
