+++ deriva/transfer/download/processors/__init__.py
@@ -8,8 +8,8 @@ from deriva.transfer.download.processors.query.bag_fetch_query_processor import 
 from deriva.transfer.download.processors.query.file_download_query_processor import FileDownloadQueryProcessor
 from deriva.transfer.download.processors.query.fasta_output_query_processor import FastaOutputQueryProcessor
 from deriva.transfer.download.processors.transform.geo_transform_processor import GeoExportTransformProcessor
-from deriva.transfer.download.processors.transform.interpolation_transform_processor import InterpolationTransformProcessor
-from deriva.transfer.download.processors.transform.interpolation_transform_processor import StrSubTransformProcessor
+from deriva.transfer.download.processors.transform.string_transform_processor import InterpolationTransformProcessor, \
+    StrSubTransformProcessor
 from deriva.transfer.download.processors.postprocess.identifier_post_processor import GlobusIdentifierPostProcessor, \
     MinidIdentifierPostProcessor
 from deriva.transfer.download.processors.postprocess.transfer_post_processor import Boto3UploadPostProcessor
