+++ deriva/transfer/download/processors/transform/string_transform_processor.py
